Rename utils to nerds_utils
