+++ producers/nmap_services_py/nmap_services_py.py
@@ -9,8 +9,8 @@ import time
 import gc
 import sys
 sys.path.append('../')
-from utils.file import save_to_json
-from utils.nerds import to_nerds
+from nerds_utils.file import save_to_json
+from nerds_utils.nerds import to_nerds
 
 logger = logging.getLogger('nmap_services_py')
 logger.setLevel(logging.INFO)
