+++ producers/raritan_snmp/raritan_snmp.py
@@ -9,8 +9,8 @@ import re
 import sys
 sys.path.append('../')
 
-from utils.file import save_to_json
-from utils import nerds as _nerds
+from nerds_utils.file import save_to_json
+from nerds_utils import nerds as _nerds
 
 logger = logging.getLogger('raritan_snmp')
 logger.setLevel(logging.WARNING)
